% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglify.R
\name{trianglify}
\alias{trianglify}
\title{Trianglify}
\usage{
trianglify(width = "100\%", height = NULL, elementId = NULL,
  cell.size = NULL, variance = NULL, seed = NULL, x.colors = NULL,
  y.colors = NULL, color.space = NULL, stroke.width = NULL,
  points = NULL)
}
\arguments{
\item{width, height}{Widget dimensions.}

\item{elementId}{Widget id.}

\item{cell.size}{Specify the size of the mesh used to generate triangles.
Larger values will result in coarser patterns, smaller values will result
in finer patterns. Note that very small values may dramatically increase
the runtime of Trianglify.}

\item{variance}{Specify the amount of randomness used when generating triangles.}

\item{seed}{Seeds the random number generator to create repeatable patterns.
When set to null, the random number will be seeded with random values from the
environment. An example usage would be passing in blog post titles as the seed
to generate unique trianglify patterns for every post on a blog that won't change
when the page reloads.}

\item{x.colors, y.colors}{Vector of colors.}

\item{color.space}{Set the color space used for generating gradients.
Supported values are rgb, hsv, hsl, hsi, lab and hcl.}

\item{stroke.width}{Specify the width of the stroke on triangle shapes in the pattern.
The default value is the ideal value for eliminating antialiasing artifacts when
rendering patterns to a canvas.}

\item{points}{A data.frame of \code{x} and \code{y} points to triangulate.}
}
\description{
Trianglify.
}
\examples{
trianglify()

}
